Move Directions to Syntax.Directions + rename
