+++ src/swarm-engine/Swarm/Game/Step/Util/Inspect.hs
@@ -15,7 +15,7 @@ import Swarm.Game.Robot
 import Swarm.Game.State
 import Swarm.Game.State.Robot
 import Swarm.Game.Universe
-import Swarm.Language.Direction
+import Swarm.Language.Syntax.Direction
 import Swarm.Util (listEnums)
 
 -- * World queries
