+++ src/swarm-lang/Swarm/Language/Syntax/Util.hs
@@ -33,10 +33,10 @@ import Data.List.NonEmpty qualified as NonEmpty
 import Data.Set qualified as S
 import Data.Tree
 import Swarm.Language.Context (Var)
+import Swarm.Language.Syntax.AST
 import Swarm.Language.Syntax.Constants
 import Swarm.Language.Syntax.Loc
 import Swarm.Language.Syntax.Pattern
-import Swarm.Language.Syntax.Type
 
 -- Setup for doctests
 
@@ -45,7 +45,7 @@ import Swarm.Language.Syntax.Type
 -- >>> import Swarm.Language.Syntax.Constants
 -- >>> import Swarm.Language.Syntax.Loc
 -- >>> import Swarm.Language.Syntax.Pattern
--- >>> import Swarm.Language.Syntax.Type
+-- >>> import Swarm.Language.Syntax.AST
 
 -- | Make an infix operation (e.g. @2 + 3@) a curried function
 --   application (e.g. @((+) 2) 3@).
