+++ src/swarm-scenario/Swarm/Game/Location.hs
@@ -56,7 +56,7 @@ import Swarm.Util qualified as Util
 -- $setup
 -- >>> import qualified Data.Map as Map
 -- >>> import Linear
--- >>> import Swarm.Language.Direction
+-- >>> import Swarm.Language.Syntax.Direction
 
 -- | A t'Location' is a pair of @(x,y)@ coordinates, both up to 32 bits.
 --   The positive x-axis points east and the positive y-axis points
