+++ src/swarm-scenario/Swarm/Game/Scenario/Topography/Structure/Assembly.hs
@@ -25,7 +25,7 @@ import Swarm.Game.Scenario.Topography.Area
 import Swarm.Game.Scenario.Topography.Navigation.Waypoint
 import Swarm.Game.Scenario.Topography.Placement
 import Swarm.Game.Scenario.Topography.Structure
-import Swarm.Language.Direction (directionJsonModifier)
+import Swarm.Language.Syntax.Direction (directionJsonModifier)
 import Swarm.Util (commaList, quote, showT)
 
 -- | Destructively overlays one direct child structure
