+++ src/swarm-tui/Swarm/TUI/View/Structure.hs
@@ -29,7 +29,7 @@ import Swarm.Game.Scenario.Topography.Structure.Recognition.Registry (foundByNam
 import Swarm.Game.Scenario.Topography.Structure.Recognition.Type
 import Swarm.Game.State
 import Swarm.Game.State.Substate
-import Swarm.Language.Direction (directionJsonModifier)
+import Swarm.Language.Syntax.Direction (directionJsonModifier)
 import Swarm.TUI.Model.Name
 import Swarm.TUI.Model.Structure
 import Swarm.TUI.View.Attribute.Attr
